[PInvoke] Prepare for 0.5.

+++ branches/libbulletc/src/btSerializer_wrap.h
@@ -18,7 +18,7 @@
 #define pSerializer_StartSerialization void*
 #define btSerializerWrapper void
 #else
-typedef btChunk* (*pSerializer_Allocate)(int size, int numElements);
+typedef btChunk* (*pSerializer_Allocate)(size_t size, int numElements);
 typedef void (*pSerializer_FinalizeChunk)(btChunk* chunk, const char* structType, int chunkCode, void* oldPtr);
 typedef const char* (*pSerializer_FindNameForPointer)(const void* ptr);
 typedef void* (*pSerializer_FindPointer)(void* oldPtr);
